Add stringAvatar helper function to Avatar and UserProfile components
