+++ frontend/src/components/molecules/Avatar.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import MuiAvatar from "@mui/material/Avatar";
 import PersonIcon from "@mui/icons-material/Person";
+import { stringAvatar } from "@/utils/helpers";
 
 interface AvatarProps {
   name: string;
@@ -8,38 +9,6 @@ interface AvatarProps {
   startDate: Date;
   email?: string;
   phone?: string;
-}
-
-function stringToColor(string: string) {
-  let hash = 0;
-  let i;
-
-  /* eslint-disable no-bitwise */
-  for (i = 0; i < string.length; i += 1) {
-    hash = string.charCodeAt(i) + ((hash << 5) - hash);
-  }
-
-  let color = "#";
-
-  for (i = 0; i < 3; i += 1) {
-    const value = (hash >> (i * 8)) & 0xff;
-    color += `00${value.toString(16)}`.slice(-2);
-  }
-  /* eslint-enable no-bitwise */
-
-  return color;
-}
-
-function stringAvatar(name: string) {
-  return {
-    sx: {
-      bgcolor: stringToColor(name),
-      height: "125px",
-      width: "125px",
-      fontSize: "75px",
-    },
-    children: `${name.split(" ")[0][0]}${name.split(" ")[1][0]}`,
-  };
 }
 
 const Avatar = ({ image, name, startDate, email, phone }: AvatarProps) => {
