+++ frontend/src/utils/helpers.ts
@@ -131,3 +131,38 @@ export const formatDateTimeToUI = (datetime: string) => {
 
   return [formattedDate, timeRange];
 };
+
+/**
+ * Generates a string avatar for the given name
+ * @param name is the name of the user
+ */
+export function stringAvatar(name: string) {
+  function stringToColor(string: string) {
+    let hash = 0;
+    let i;
+
+    /* eslint-disable no-bitwise */
+    for (i = 0; i < string.length; i += 1) {
+      hash = string.charCodeAt(i) + ((hash << 5) - hash);
+    }
+
+    let color = "#";
+
+    for (i = 0; i < 3; i += 1) {
+      const value = (hash >> (i * 8)) & 0xff;
+      color += `00${value.toString(16)}`.slice(-2);
+    }
+    /* eslint-enable no-bitwise */
+
+    return color;
+  }
+  return {
+    sx: {
+      bgcolor: stringToColor(name),
+      height: "125px",
+      width: "125px",
+      fontSize: "75px",
+    },
+    children: `${name.split(" ")[0][0]}${name.split(" ")[1][0]}`,
+  };
+}
